Renamed GaKCo Gakco; ask user for m instead of k
