+++ src/Gakco.h
